Changed implementation for Task2 and Task3

+++ src/main/java/edu/hw2/Task2/Rectangle.java
@@ -17,7 +17,7 @@ public class Rectangle {
         return new Rectangle(this.width, height);
     }
 
-    public double area() {
+    public int area() {
         return width * height;
     }
 }
